use double type instead of float

+++ src/algo/AnchorFinder.cpp
@@ -168,7 +168,7 @@ void AnchorFinder::run_impl() const {
     if (std::log(length_sum) / std::log(4) > anchor_size) {
         length_sum = std::pow(double(4), double(anchor_size));
     }
-    float error_prob = 1.0 / length_sum;
+    double error_prob = 1.0 / length_sum;
     std::set<size_t> possible_anchors;
     {
         BloomFilter filter(length_sum, error_prob);
